pyproject.toml
[tool.poetry]
name = "drf-homework"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
pylint = "^3.1.0"
isort = "^5.13.2"
psycopg2 = "^2.9.9"
pylint-django = "^2.5.5"
django-stubs = "^4.2.7"
pylance = "^0.10.13"
flake8 = "^7.0.0"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
